modified:   infraninja/security/common/reboot_system.py

+++ infraninja/security/common/reboot_system.py
@@ -1,10 +1,11 @@
-
 from pyinfra.operations import server
 from pyinfra.api import deploy
 
+# When calling this deploy function, you can pass a boolean value to the need_reboot parameter
 @deploy("Reboot the system")
-def reboot_system():
-    server.reboot(
-        name="Reboot the system",
-        delay=10,  # Delay in seconds before reboot
-    )
+def reboot_system(need_reboot=None):
+    if need_reboot is None or need_reboot:
+        server.reboot(
+            name="Reboot the system",
+            delay=10,  # Delay in seconds before reboot
+        )
